chore: move all wallets into individual files
