+++ src/lib/data/wallets/walletImports.js
@@ -1,0 +1,34 @@
+import contributionWallets from './contributionWallets.json';
+import conversionServiceWallets from './conversionServiceWallets.json';
+import depositWallets from './depositWallets.json';
+import ignoreWallets from './ignoreWallets.json';
+import investmentWallet  from './investmentWallet.json';
+import investmentWallets from './investmentWallets.json';
+import legacyWallets from './legacyWallets.json';
+import memberWallets from './memberWallets.json';
+import membershipWallets from './membershipWallets.json';
+import stableCoinContracts from './stableCoinContracts.json';
+import teamWallets from './teamWallets.json';
+import tokenContractAddresses from './tokenContractAddresses.json';
+import vcWallets from './vcWallets.json';
+import velaWallets from './velaWallets.json';
+
+export default {
+    investmentWallet: investmentWallet.investmentWallet,
+
+    wallets: {
+        contributionWallets,
+        conversionServiceWallets,
+        depositWallets,
+        ignoreWallets,
+        investmentWallets,
+        legacyWallets,
+        memberWallets,
+        membershipWallets,
+        stableCoinContracts,
+        teamWallets,
+        tokenContractAddresses,
+        vcWallets,
+        velaWallets
+    }
+};
